+++ CHANGELOG.md
@@ -68,3 +68,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Menambahkan CRUD pada modul fine, fine-payment, transaction, stack, dan shelf
+
+### Change
+- Menyusun ulang arsitektur pada modul controllers, services, repositories, dan routes